+++ lib/public/components/qcFlags/qcFlagCreationPanelLink.js
@@ -85,35 +85,3 @@ export const qcFlagCreationPanelLink = (
         },
     });
 };
-
-/**
- * Render link to QC flag overview page
- *
- * @param {*} content content to display
- * @param {number} [scope.dataPassId] data pass id - exclusive with simulationPassId parameter
- * @param {number} [scope.simulationPassId] simulation pass id - exclusive with dataPassId parameter
- * @param {number} scope.runNumber run number
- * @param {number} scope.dplDetectorId DPL detector id
- * @param {object} [attributes] attributes @see frontLink attributes parameter
- * @return {Component} link
- */
-export const qcFlagOverviewPanelLink = (content, { dataPassId, simulationPassId, runNumber, dplDetectorId }, attributes = {}) => {
-    let options = {};
-    let pageName = '';
-    if (dataPassId) {
-        options = { dataPassId };
-        pageName = 'qc-flags-for-data-pass';
-    } else if (simulationPassId) {
-        options = { simulationPassId };
-        pageName = 'qc-flags-for-simulation-pass';
-    } else {
-        options = {};
-        pageName = 'synchronous-qc-flags';
-    }
-    return frontLink(
-        content,
-        pageName,
-        { runNumber, dplDetectorId, ...options },
-        { title: 'QC flags overview', ...attributes },
-    );
-};
