+++ lib/public/views/QcFlags/ForDataPass/QcFlagsForDataPassOverviewPage.js
@@ -17,7 +17,7 @@ import { frontLink } from '../../../components/common/navigation/frontLink.js';
 import { estimateDisplayableRowsCount } from '../../../utilities/estimateDisplayableRowsCount.js';
 import { table } from '../../../components/common/table/table.js';
 import { qcFlagsActiveColumns } from '../ActiveColumns/qcFlagsActiveColumns.js';
-import { qcFlagCreationPanelLink } from '../../../components/qcFlags/qcFlagsPagesButtons.js';
+import { qcFlagCreationPanelLink } from '../../../components/qcFlags/qcFlagCreationPanelLink.js';
 import { qcFlagsChartComponent } from '../qcFlagsVisualization/qcFlagsChartComponent.js';
 import { mergeRemoteData } from '../../../utilities/mergeRemoteData.js';
 import { qcFlagsBreadcrumbs } from '../../../components/qcFlags/qcFlagsBreadcrumbs.js';
