+++ lib/public/views/QcFlags/ForSimulationPass/QcFlagsForSimulationPassOverviewPage.js
@@ -18,7 +18,7 @@ import { estimateDisplayableRowsCount } from '../../../utilities/estimateDisplay
 import { qcFlagsActiveColumns } from '../ActiveColumns/qcFlagsActiveColumns.js';
 import { table } from '../../../components/common/table/table.js';
 import { paginationComponent } from '../../../components/Pagination/paginationComponent.js';
-import { qcFlagCreationPanelLink } from '../../../components/qcFlags/qcFlagsPagesButtons.js';
+import { qcFlagCreationPanelLink } from '../../../components/qcFlags/qcFlagCreationPanelLink.js';
 import { qcFlagsBreadcrumbs } from '../../../components/qcFlags/qcFlagsBreadcrumbs.js';
 
 const TABLEROW_HEIGHT = 35;
