+++ lib/public/views/Runs/ActiveColumns/runDetectorsAsyncQcActiveColumns.js
@@ -11,11 +11,12 @@
  * or submit itself to any jurisdiction.
  */
 import { h, iconBan, iconX, RemoteData } from '/js/src/index.js';
-import { qcFlagCreationPanelLink, qcFlagOverviewPanelLink } from '../../../components/qcFlags/qcFlagsPagesButtons.js';
+import { qcFlagCreationPanelLink } from '../../../components/qcFlags/qcFlagCreationPanelLink.js';
 import { tooltip } from '../../../components/common/popover/tooltip.js';
 import { isRunNotSubjectToQc } from '../../../components/qcFlags/isRunNotSubjectToQc.js';
 import { getQcSummaryDisplay } from './getQcSummaryDisplay.js';
 import { getRunQcExclusionReason } from '../../../components/qcFlags/getRunQcExclusionReason.js';
+import { qcFlagOverviewPanelLink } from '../../../components/qcFlags/qcFlagOverviewPanelLink.js';
 
 /**
  * Render QC summary for given run and detector
