+++ lib/public/views/Runs/ActiveColumns/runDetectorsSyncQcActiveColumns.js
@@ -12,7 +12,7 @@
  */
 import { h, iconBan } from '/js/src/index.js';
 import { getQcSummaryDisplay } from './getQcSummaryDisplay.js';
-import { qcFlagOverviewPanelLink } from '../../../components/qcFlags/qcFlagsPagesButtons.js';
+import { qcFlagOverviewPanelLink } from '../../../components/qcFlags/qcFlagOverviewPanelLink.js';
 
 /**
  * Factory for detectors related active columns for synchronous QC flags configuration
